Move Default configuration into core module.
