+++ app/build.gradle
@@ -84,6 +84,7 @@ android {
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
     compile project(':bankdroid-legacy')
+    compile project(':bankdroid-core')
     compile 'com.jakewharton:butterknife:6.1.0'
     compile "com.android.support:appcompat-v7:22.0.+"
     compile 'com.google.collections:google-collections:1.0'
