+++ app/src/main/java/com/liato/bankdroid/BankEditActivity.java
@@ -26,10 +26,10 @@ import com.liato.bankdroid.banking.Account;
 import com.liato.bankdroid.banking.Bank;
 import com.liato.bankdroid.banking.BankChoice;
 import com.liato.bankdroid.banking.BankFactory;
-import com.liato.bankdroid.banking.DefaultConnectionConfiguration;
 import com.liato.bankdroid.banking.exceptions.BankChoiceException;
 import com.liato.bankdroid.banking.exceptions.BankException;
 import com.liato.bankdroid.banking.exceptions.LoginException;
+import com.liato.bankdroid.configuration.DefaultConnectionConfiguration;
 import com.liato.bankdroid.db.DBAdapter;
 import com.liato.bankdroid.utils.FieldTypeMapper;
 import com.liato.bankdroid.utils.NetworkUtils;
