+++ bankdroid-core/build.gradle
@@ -1,0 +1,8 @@
+apply plugin: 'java'
+
+sourceCompatibility = JavaVersion.VERSION_1_7
+targetCompatibility = JavaVersion.VERSION_1_7
+
+dependencies {
+	compile project(':bankdroid-interface')
+}
