+++ bankdroid-core/src/main/resources/i18n/application.properties
