+++ bankdroid-core/src/main/resources/i18n/application_sv_SE.properties
